requirements.txt
scikit-learn>=1.0
pandas
numpy
matplotlib
seaborn
joblib
datasets
scipy
joblib
python-telegram-bot>=13.15
python-dotenv
datasets
pandas
scikit-learn
joblib
pyarrow
python-telegram-bot
